rename/att geoLoc
