+++ src/geoLoc/dto/update-geoLoc.input.ts
@@ -1,5 +1,5 @@
 import { InputType, Field, Float, PartialType } from '@nestjs/graphql';
-import { CreateLocalizationInput } from './create-localization.input';
+import { CreateLocalizationInput } from './create-geoLoc.input';
 
 @InputType()
 export class UpdateLocalizationInput extends PartialType(CreateLocalizationInput) {
