+++ src/geoLoc/geoLoc.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { LocalizationService } from './localization.service';
-import { LocalizationResolver } from './localization.resolver';
+import { LocalizationService } from './geoLoc.service';
+import { LocalizationResolver } from './geoLoc.resolver';
 import { PrismaService } from 'src/database/service/prisma.service';
 
 @Module({
